backend/requirements.txt
django==2.0.5
psycopg2==2.7.5
djangorestframework==3.8.2

